+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@
 - `[Bar Chart]` Added support to ellipsis for yaxis labels. ([#3702](https://github.com/infor-design/enterprise/issues/3702))
 - `[Contextmenu]` Added support for shortcut display in menus. ([#3490](https://github.com/infor-design/enterprise/issues/3490))
 - `[Datepicker]` Added support for custom api callback to disable passed dates and to disable dates by years. ([#3462](https://github.com/infor-design/enterprise/issues/3462))
+- `[Datagrid]` Added and fixed up datagrid grouping aggregators. There is now aggregators for avg, count, list, max, min and sum. In addition null and undefined data will not cause issues. ([#3752](https://github.com/infor-design/enterprise/issues/3752))
 - `[Error Page]` Added a new example showing a static error page. For example for a 404 page or generic error. ([#281](https://github.com/infor-design/design-system/issues/281))
 - `[FileUploadAdvanced]` Added support to api settings `maxFiles` to limit number of uploads. ([#3512](https://github.com/infor-design/enterprise/issues/3512))
 - `[FileUploadAdvanced]` Added support to fire event `fileremoved` for attached file removed. ([#3548](https://github.com/infor-design/enterprise/issues/3548))